requirements_dev.txt
bumpversion==0.5.3
wheel==0.26.0
watchdog==0.8.3
flake8==2.5.1
tox==2.3.1
coverage==4.0.3
Sphinx==1.3.3

pytest==2.8.5
pytest-django==2.9.1
django_compressor==2.0
hashids==1.1.0
